refactor(useCartItem): error 처리 로직 제거
